add flushoncommit to btrfs mountopts
By default, btrfs doesn't necessarily do a sync() before it takes a
snapshot. So, you can have a situation where you do a write, take a
snapshot of the directory containing that write, and your snapshot doesn't
contain that write. This is the setting to fix that. Why it defaults to off
is beyond me.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>