add support for unprivileged stacker
The way we're doing this is sort of a hack, but also sort of not: we
re-exec stacker in a user namespace with appropriate mappings, and then
just let it run normally. It is a hack because ideally we wouldn't have to
re-exec, and every bit of code would perfectly understand user namespaces.

However, it seems like quite a bit less of a hack because we don't actually
have to teach all the code about user namespaces (except for example the
convenience chmod at the end); it all just looks exactly the same as it
does for the root case. That actually seems quite nice.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>