don't hard code host mapped uid
We shouldn't hard code this, because if users have > 100k uids mapped, it
will fail. In order to not hard code this, we need to fiddle with the
internal flag a bit and make it take the argument to let the
inside-user-ns task know where to chmod things when it's done. But since
the flag was internal, this should be ok :)

Fixes #2
Signed-off-by: Tycho Andersen <tycho@tycho.ws>