container: bind mount /sys instead of using autodev sys:mixed
The problem here is that the kernel refuses to let you mount sys if you
don't own the netns that you're currently in. Since we're using host
networking, we don't own our netns, and this is always refused.

Since we don't know if any packages will actually require sys, let's just
bind mount it instead, even if it's mostly unusable.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>