rearrange the way we handle user namespaces
Previously, we were re-execing ourselves inside a user namespace with
lxc-usernsexec and then running the entirety of the stacker code. The
reason was so that we didn't have to annotate individual parts of the code
with RunInUserns, or explicitly re-exec in these situations.

However, this caused various problems, not in the least of which is that
lxc-usernsexec didn't really set up a full container, so the run containers
that were spawned by stacker required lots of little hacks in liblxc and
elsewhere to work. Also, it was confusing.

Instead, let's switch to running the stacker in the init user ns, and
actually do annotate the various points in the code which need to be exec'd
in a userns. It turns out this wasn't very complicated, since we're already
shelling out to umoci to do most/all of the unpacking, and starting a
container to do grab.

This has the further advantage that it no longer violates the principle of
least surprise.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>