cache: also depend on imports changing
Previously, stacker only looked at the layer recipe itself, so if e.g. the
path to copy in hadn't changed, but the contents had, we didn't rebuild.

Now, we inspect the import contents as well. Unfortunately we can't just
hash the file, since we also support importing directories. Instead, we
keep a cache of the mtree output. Also unfortunately, since the mtree
output doesn't serialize to json very well, we keep a base64 encoded
version of the binary string. There's probably a better way to do this, and
we can bump to v3 if it seems important :)

Signed-off-by: Tycho Andersen <tycho@tycho.ws>