cache: actually tag the descriptor with the current name
Update the descriptor this tag points to in case someone updated the
descriptor out of band, and remove comment suggesting we do this.

We could potentially re-build if there's an error here, but a cache pruning
commit is coming shortly, so in principle we should only get this error if
there actually is a permissions problem.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>