actually respect ^C
So it turns out the problem here is that liblxc sets ISIG on the
controlling terminal, which ignores ^C characters and doesn't generate
sigterm. Of course, it really wants to set this on the container's
terminal, not the host's terminal, but because we didn't do setsid(), we
didn't unset the controlling terminal, and so escape sequences were ignored
on the host.

Adding setsid() causes its own problems though, and disconnects stdout/err.
So we need to set those to a pipe, and have a goroutine copy stuff from the
pipe to the real stdout. But only in some cases, because we still do want
the interactive mode for --on-run-failure.

This patch handles all of that.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>