apply: make layer merging smarter
A fairly large change, but there it is. The idea here is to just insert the
layers as-is into the output image, so that some day we could swap them
out with additional metadata and tooling help.

This commit also disables the "merging" code. The problem is that the umoci
metadata is an mtree file of things as they are on disk, and if we merge
things inline, we'll generate the mtree file with the on-disk changes, and
the final merge will be lost.

We could solve this by keeping unmerged changes on the side, or appending
mtree files (if possible), etc. But in any case, it doesn't work now, and
the intelligent layer merging code is more important, so that supercedes
things.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>