import: allow importing broken links
It is reasonable to want to import broken links, so let's allow that. Note
that this doesn't *quite* allow importing broken links: the caching bits
still use Stat() in various places, so if the top level thing you're trying
to import is a broken link, it will still barf. However, this seems like a
pathological case and so I'm going to ignore it.

It turns out that's slightly hard to get right, because we previously
allowed symlinks other files and treated them as that actual file, so we'd
have to have some extra special cases for backward compatibility.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>