squashfs layers: don't exclude everything from generation
The problem here was that directories like /etc and /usr would always be
the same, because we rarely modify them, and thus mtree would dutifully
generate a Same entry for them, and we would dump it into the list of
things to exclude.

Then, mksquashfs would see that nothing matches the list of things to
exclude, and so it wouldn't do anything.

To fix this, let's add our own little excludePaths helper, which keeps
around the lists of paths to include and exclude, and makes sure they don't
conflict.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>