apply: remove mtree file before trying to generate it
GenerateBundleManifest() fails if the mtree file already exists, but in
some cases (e.g. the restore-from-snapshot case that will be introduced in
the next patch) it may already exist. So, let's delete it so we can
re-generate it.

In particular: it won't be the same, since apply modified stuff, so we
can't just skip this step and re-use it.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>