unlade: extract to .working first
Everything else in the code extracts to .working and then snapshots to
wherever they want the real data to go. unlade doesn't do that (it's never
going to run a container on top of the stuff), but for consistency and so
we can keep a hack that's in the next patch, let's make unlade do it this
way.

This has the added advantage that now these snapshots are readonly, as the
snapshots from the other generation methods are. Since we weren't
snapshotting, just Create()ing under the tag name, snapshots weren't
readonly before.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>