chroot: always assume _working if nothing is specified
This is really what people meant: typically we use stacker chroot after a
failed build to examine what went wrong. The code before worked because of
the way people set up their stacker files, but that was mostly a
coincidence.

In particular, with all the new snapshotting intermediate layers code, lots
of extra dirs will be generated in roots, and it's unlikley that we'll
guess the right one.

Of course, people can still pass what they want if they want to do
something manually.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>