base: rewrite history when generating squashfs layers
The problem here is that some tools (containers/image, probably others)
assume that the number of non-empty entries in the history for image config
is == to the number of layers. However, when we re-generate the layers to
be a single squashfs layer, we've violated this assumption. So let's
regenerate the history too.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>