intermediate base snapshots: fix last+last extraction
The problem here was rooted in the fact that we didn't snapshot the
last unpacked layer with all the metadata in it. The fix is to do this
snapshot again after everything has been unpacked, and do a more
comprehensive metadata clean when we're about to unpack something new.

Gory details from the testcase are:

t1 would create a last layer with no umoci.json/mtree file.
t2 would generate the mtree/umoci.json for the next layer up correctly
since it is generated via prepareUmociMetadata(); then it would remove the
mtree since it was going to be regenerated, but not the umoci.json. but
since we didn't snapshot the layer with the generated umoci metadata...
t3 would look, see an umoci.json but no mtree, and barf

This points out another optimization: if we're going to extract over the
top of what we're doing, we *don't* need to prepareUmociMetadata(). That
optimization is in the next patch.

Closes #40

Signed-off-by: Tycho Andersen <tycho@tycho.ws>