umoci unpack: only generate metadata if we're not going to do it later
In the intermediate layer case, we may generate the umoci metadata after
the last layer is extracted; any work we do in the middle regenerating the
metadata is wasted. So let's only do this step if the layer we're matching
the hash of is the last layer in the manifest.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>