Support shells/languages other than /bin/sh in run section.
If the 'run' section starts with a '#!' such as '#!/bin/bash' or
'#!/usr/bin/python', then assume that the user wanted that interpreter.

I have to admit that my primary motivation for this is actually getting
rid of the '-x' and '-e' flags that stacker invokes a run section with.
But just dropping the '-e' at least would likely break existing
run values.  So instead, add support for whatever the user wants.

run: |
  #!/bin/sh
  echo "Horay, your logs do not show this line twice due to set -x"
  echo "And if this command exited non-zero the next line would run!"
  echo "Read http://mywiki.wooledge.org/BashFAQ/105"