import: consider URL change
When computing whether or not two *layers* were the same, we correctly took
the URL into account and forced a rebuild. However, when computing whether
two *imports* were the same and needed to be re-downloaded, we did not.
Instead, we just looked to see if a file with that particular basename was
present or not.

This problem can also occur with local paths (i.e. not just URLs).

So, let's look at the old cache entry and see if any of the import URLs for
a particular directive have changed and if they have, just delete that from
the cache before we do imports.

Perhaps this should be a part of the caching code itself, that when there
is an explicit cache miss because of layer import changes, it should delete
the stuff from the cache. It also seems weird to have a cache that mutates
the disk, though. So I'll leave it like this for now.

Fixes #55 take two :)

Signed-off-by: Tycho Andersen <tycho@tycho.ws>