base: leave the output tag around when importing built types
The semantics of e.g. getContainersImageType() (but indeed, should be the
semantics of all sub-implementations of GetBaseLayer()) are that it imports
the image type and leaves the output name as the image tag in the output
OCI image. getBuilt() didn't do this, so let's fix it so it does.

The squashfs layer generation backend relies on this since it updates the
existing tag, so if it doesn't exist, things fail. Let's fix this and add a
test to make sure it continues to work.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>