squashfs: don't record modified/extra files as missing
The missing list is files that we'll delete after we're done generating the
layer; this is intended to be a list of the places we did a mknod. However,
a copypasta error from 26f50e895ecd30fc89e448699712f53e07c49dac meant that
we also deleted files that were new from the layer, which was not helpful,
and basically meant we deleted everything after generating a squashfs layer
from it.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>