base: fix incorrect importing of base layers
The goal of the import here is to get whatever OCI stuff that already
exsits into the output image, so that we can add the new stuff on top of
this.

We were doing this incorrectly: we don't want to go to the first non-built
layer, in fact we want the closest reasonable layer to what's there now. So
instead of doing a recursive look through the layer definitions, let's look
at the base layer's output, if it doesn't exist, look for its input, and if
that doesn't exist, give up and put the whole rootfs in the next OCI
output.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>