Remove automatic tagging of uploaded layers with commit-<id>
The commit-<id> tag was automatically added to all the newly published layers, and the user had no control over it.
This could have had unintended consequences, if the destination server where the tag was uploaded already
contained a layer with the same name and tag.
The tag would point to the new layer and the reference to the older layer would be lost.

It is better to have the user explicity mention in the command line all the tags he wants to use, and push no others.