Revert "caching.bats: use alpine"
This reverts commit 0f5ac4996d257e7d3541eb6475809a35de11e076.

Unfortunately, the ubuntu packages for lxc provide an lxc-init that is not
statically linked, and is also the default. The default centos image has
libcap, so this lxc-init works, but the alpine image does not.

Since we run our CI in ubuntu, we get a CI failure here.

Let's fix it by just using centos:latest for now. Hopefully at some point
the lxc packages will be fixed to provide a static init as the default, and
this will work.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>