support upriv squashfs
When unpacking and creating squashfs images, we need to do more than
we were in a user namespace,  This is because the rootfs may have
files and dirs which are not world readable, so the unpriv user cannot
read them without having privilege (in a user namespace) over the owning
uid.

Create a new 'squashfs' subcommand to unpack and repack.

Also, the squashfs path was always using a chardev to create whiteouts.
You can't do that in a user namespace.  So if that fails, fall back on
creating .wh.$path.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>