lib: refactor container object usage
The run.go file has a thin wrapper around container which makes it awkward
for many users.

Instead, let's move most of this functionality into container.go, with the
stacker-specific stuff living in its own method.

This will be useful for the next commit which implements the configure-only
step for stacker, since it separates the "configure" and "run" logic. It
has the added benefit of dropping a silly abstraction :)

Signed-off-by: Tycho Andersen <tycho@tycho.ws>