build: split base layer acquiring and extraction
GetBaseLayer really did a lot more work than just getting the base layer;
it got the base layer, but it also initialized the umoci metadata on the
output filesystem (which was created elsewhere), and copied stuff to the
output.

In order to check whether the new base layer is different from the old one
for caching purposes, we need to split this operation into at least two
distinct parts: getting the base layer, and doing all the rootfs setup.
This commit does that. SetupRootfs() still does the second two steps, which
may or may not be good. In any case, with this split, we can enable
checking of the base layer in the cache.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>