cache: drop LookupWithoutChecks()
Using LookupWithoutChecks() is essentially always incorrect. In
getBaseHash(), we really want to do the checks for the base layer, because
if the base layer generates a cache miss, we shouldn't generate a cache hit
later.

In the case of the one in the import cleaning bits, we can just drop it,
since the subsequent cleaning/import code should handle any mismatches.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>