cache: better indication of failures in Lookup()
Instead of swallowing errors for cache misses, let's surface them. This
gives us a better indication of when some lookup failed or when there was
some other error while doing the lookup.

This also means that the Lookup() function should either have 1. an error
returned or 2. a logging statement indicating why the lookup failed, which
is a nice side effect.

Might be cool to hoist this out into a type to make it type safe :)

Signed-off-by: Tycho Andersen <tycho@tycho.ws>