cache: partially revert LookupWithoutChecks()
This is a partial spiritual revert of
44953b24d5de5e1153b172f6317e3416e17d0b3d.

Turns out we *do* want some of this behavior: when cleaning old imports, we
want to compare the new URL against the old URL, so we do want to lookup
things, but without doing all the hashing and whatnot associated with a
"real cache lookup".

However, we don't need the proxy LookupWithoutChecks() method, we can just
use the cache object directly, since it's public.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>