container: don't resolve idmap on init()
As @smoser points out in #88, Warn()s are ugly.

It seems fairly impolite to go around doing a bunch of work during init().
It slows down everything for someone who may never use this code, and it's
hard to report errors.

Instead of reporting errors, we had a bunch of warnings (which were
themselves hard to get right and were the source of obfuscation of errors.
See bd8554d45459 ("Only print an error if there was an error") or
e9ebfc7de64d ("add more logging when uidmap parsing fails").

Instead, let's just treat this as normal code, because it is: parse the
files when they're needed, and return an error if that fails. This lets us
avoid ugly Warn()s in favor of actual failure when things fail.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>