tests: prepare for parallel execution
This cleans up various file-global pieces of state in the tests, and makes
them all run in their own independent directory.

It also adds the necessary bits to run the tests in parallel. However, it
doesn't do that by default, since there seems to be a mysterious bug caused
by enabling it :). But hopefully we can solve that soon.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>