storage: retry attaching the btrfs loop device
We can race with ourselves when attaching the loop device, so let's retry
a few times to be sure we really get it.

Note that we don't retry very much -- 100 ms * 10 = 1 second maximum,
because the window is pretty small and this seems to be enough even in
large parallel (32 threads) cases.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>