clean: do more complaining when we can't clean up correctly
In some cases, the cleanup code fails because:

~# btrfs property set -ts debug/rootfs/var/lib/machines ro false
~# echo $?
0
~# btrfs property list -ts debug/rootfs/var/lib/machines/
ro                  Set/get read-only flag of subvolume.
~# btrfs subvolume delete -c debug/rootfs/var/lib/machines
Delete subvolume (commit): '/root/c3/build/roots/debug/rootfs/var/lib/machines'
ERROR: cannot delete '/root/c3/build/roots/debug/rootfs/var/lib/machines': Read-only file system

An strace of the "set readonly false" command above shows:

openat(AT_FDCWD, "debug/rootfs/var/lib/machines", O_RDONLY) = 3
ioctl(3, BTRFS_IOC_SUBVOL_GETFLAGS, 0)  = 0
ioctl(3, BTRFS_IOC_SUBVOL_SETFLAGS, 0)  = 0
close(3)                                = 0

i.e. the kernel thinks that it set the ro property to false correctly.

The next step stacker does is just try to unmount it:

umount: roots: not mounted.

But that leaves us in an even weirder state: the fs is not mounted
according to /proc/self/mountinfo, but it the files are showing up, and you
can query subvolumes and their properties just fine. lsof shows nothing is
open on these btrfs subvolumes, so it's not being pinned because of that.

In any case, in light of all this, let's at least exit(1) and print a
message about btrfs being wedged, so people maybe know what's going on.

The above was on a 5.3.0 kernel, and we've seen versions of this issue for
years, so it seems unlikely that it'll get fixed any time soon :(

Note that the above was in a nested container, and I've not personally seen
it on a host, so it seems related to nested namespaces somehow, maybe?

Signed-off-by: Tycho Andersen <tycho@tycho.ws>