storage, btrfs: mark subvolumes writable in reverse order
When we have nested subvolumes, e.g. roots/test/var/lib/machines (where
test is a subvolume, as is machines), we need to un-mark them in reverse
order to how we delete them, since btrfs refuses to let us mark a subvolume
writable if its parent is writable.

I'm not really sure why we started hitting this all of the sudden; systemd
has long been making var/lib/machines a subvolume if it's on btrfs. It
might have something to do with the _working container removal patch, but
even that I don't really understand, since we would have previously deleted
old snapshots when re-building stuff. So presumably we should have hit this
a bunch more.

Once we started looking at errors in subvolume deletion, this manifested
almost all the time when building a centos image.

Finally, this is really only a "problem" in the non-loopback case, because
in the loopback case we'll just fail to delete the subvolumes, but still
unmount successfully and delete btrfs.loop. Since "most" of the use of
stacker (and all of the tests) use the loopback case, we never noticed.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>