build: fix shell detection in busybox images
busybox uses an absolute symlink to /bin/busybox as /bin/sh, so this stat
was pointing at the host's /bin/busybox, which most hosts won't have.

Instead, let's just do an lstat() and keep going if that works. Things can
still fail as mentioned in the comment, but "make it idiot proof and
someone will make a better idiot"; er, s/user/idiot :)

Signed-off-by: Tycho Andersen <tycho@tycho.ws>