storage: add a new overlayfs backend
This commit adds a new storage backend based on overlayfs.

We spend most of our time during builds in 1. layer extraction and 2. mtree
generation. If we use overlayfs, we can parallelize layer extraction, thus
speeding that part up as much as possible. Using overlay's upperdir=
mechanism, we can evaluate what delta a particular run: section generated
by simply walking the filesystem.

Based on this, we introduce a new storage backend based on overlayfs. It
works by extracting all containers/image style layers in parallel to
roots/$shasum/overlay, and then generating an overlayfs mount for all of
these together. We write a bit of metadata (what layers and built type
layers) to the directory so that we can reproduce the mounts if necessary,
or generate the right output when using a subsequent build.

Snapshots are really just an accumulation of these kinds of layers, since
we don't actually use the copy on write functionality of btrfs (indeed, if
something lower in the stack changes, we need to regenerate everything
above it, so any CoW usage is/was a bug).

To use this unprivileged, people need two kernel patches, one in tree in
v5.8, and one out of tree one (but that is in the ubuntu kernel). See
overlay/overlay.go for more details.

Finally, we still default to the old btrfs style of doing things, since
this is experimental.

There's still some stuff TODO to make this better faster stronger, see the
comment at the top of overlay/overlay.go.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>