run: generate a better shell warning in unespected stat return
If this fails with IsNotExist() type errors, that's fine and what we
expect. If it fails for some other reason, we should really render that
error as well.

It is still a failure, because we won't be able to run the shell, it's just
better to know what the actual failure is too.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>