btrfs: do intermediate snapshot mtree generation in a subprocess
f6668679f8eb ("use rootless fseval everywhere") is not quite correct. When
we do an mtree generation from the parent process, we do it as the real
uid of the person executing stacker, not as root inside the container.

Then later, when someone bases something on that layer and does a repack
from a properly user namespaced process, all the uids on every file look
different. This causes the repack to include every file (including the file
contents, not just the metadata fixups, because yay OCIv1 :) again, thus
doubling the size of the image.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>