base: move btrfs-specific layer-translation into btrfs
When going from tar->squashfs, we need to convert at some point. The way we
do this in the btrfs backend is different than the way we'll do it in the
overlay backend, since we can do it in overlay on a per-layer basis, but
there's not a nice way to do it like that in btrfs.

So, let's push this code into the btrfs backend.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>