build: allow for multiple outputs
Previous to this, if you want to build two different layer outputs, you'd
have to do a build twice, doing all the imports and decompressing twice,
doing all the import:s twice, doing the run:s twice, etc. This is obviously
slow.

Instead, let's enable building multiple output types at the same time.
Right now, this only works for the overlay backend.

The way it works is by generating all the layer type outputs at the same
time during the Repack() stage, which we can do now because of the previous
work moving the output generation to the Repack() stage. The basic premise
is:

for each layer:
    for each layer-type:
        generate the layer

Note that this changes how we do output slightly: even in the single layer
build case, --layer-type squashfs now outputs a layer as foo-squahsfs, not
just foo. In a multiple layer case, the tar layer will always be output as
the original tag, and -squashfs will always be appended for squashfs
layers.

While this is a divergence from how things previously worked, it is
essentially how we were using it before by building and tagging a -squashfs
layer, so it seems reasonable.

Another approach would be to add another annotation that indicates the
layer type, and emit the same tag twice along with this new annotation that
disambiguates the layer types. However, most of the other tooling (umoci,
skopeo) would need to support this, so we do this for now. But perhaps it
will change in the future, you have been warned :)

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>