publish: support multiple --layer-types
Now that we can *build* multiple layer types, it is reasonable to want to
publish them as well.

While we could run publish twice, with just the right arguments, I think it
makes more sense to just have --layer-type supported as an argument to
publish, and do the same math on layer types.

This way, if we ever come up with a better to do --layer-type output,
people won't have to go change their scripts everywhere.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>