Support use of squashtool for extraction.
Unsquashfs just does not operate really well when it is not root.
It sometimes ignores extraction errors, and other times fails.
squashtool intends to behave better when unprivileged.

If unprivileged and user has squashtool in their path, then
squashtool will be used for extraction rather than unsquashfs.

By default:
  - as root, behave as before.
  - as non-root: use squashtool if available, else unsquashfs.

The decision of whether or not the user is operating as root can
be affected by STACKER_UNPRIV in the environment.
STACKER_UNPRIV can have three values:
 * "" (or not set in env): unprivileged unless euid == 0
 * true: unprivileged (non-root)
 * false: privileged (root)