overlay: unpack the right layer
The problem here is that we capture in a closure the iteration variable,
whose value changes with each iteration.

Let's move to assigning the parts we really care about (i.e. layer.Digest)
to a new heap location which is allocated on each iteration of the loop, so
it stays the same value when captured in the closure.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>