build: don't Mount()/Detach() storage on each iteration
If we're building multiple stackerfiles (or stacker files with
prerequisites, etc.) let's not re-configure the storage every time, but do
it once at the beginning and detach it when we're all done.

Note that this does fix a bug: people using stacker:// imports with the
overlay backend wouldn't see these, since the Detach() in the middle would
unmount the rootfs dir, and nothing remounts it. We should definitely
remount anything that looks like it might be used to prevent this bug
between stacker invocations. Add a TODO for that.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>