tests: remove the right thing in intermediate snapshot test
Since we changed the outputs to be $foo-squashfs, we need to change these
as well.

This didn't manifest before a4274f1280fd ("build: correctly detect cached
layers for multiple types") since we had a bug where we would always
rebuild. But now that we don't always rebuild, we see this bug!

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>