import: handle readonly parent directories
in the case where a parent directory of an import is read-only, things may
fail in the unprivileged case because non-root cannot rm -rf a file that's
-w.

To fix this, we chmod +w the parent directory, and chmod it back at the
end.

Note that we can't defer the chmod back in chmodParentAndRemove(), since
the subsequent copy will also not be able to overwrite the file, since the
dir will be -w again.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>