btrfs: require squashtool to extract squashfs images
It is always incorrect to use unsquashfs to extract things in btrfs mode,
since btrfs requires an unlink() instead of mknod c 0 0, which is how we
store whiteouts in squashfs images.

Indeed, unprivileged executions can't mknod anyways on kernels < 5.8, so
it's both incorrect and won't even work in many cases.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>