unpriv-setup: do something reasonable for overlay mode
for overlay, we definitely don't want to mount a btrfs loopback, as that'll
slow things down making it all go through the loopback device.

However, we do still need the uid setup, so it's not entirely a no-op.

It would be nice to check that the unprivileged user can actually create
whiteouts and mount overlay, but golang doesn't really support setuid()
(it's single threaded on linux) which makes that kind of hard. So we'll
just wait for the check on use to fail, which isn't awful.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>