container: run root containers as root
6a0b2b601ba4 ("container: rearrange the way we resolve idmaps") changed
this so it always returns a non-nil map if root has subuids. However, we
don't have code to fix up the stacker dir to have the right acls (which
breaks things on travis), so instead if we're root, just run containers as
root.

Probably the right way to fix this is to detect SUDO_USER or something.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>