storage: switch storage types
Previously, we didn't store the storage type, so we couldn't reason about
it. Let's store the storage type so we can use it to reason about if we
should clean the previous storage or not.

Additionally, to support upgrade from when we didn't store storage types,
let's add a heuristic for detecting previous storage.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>