test suite: don't download 3459827345234 copies of centos:latest
The docker hub introduced new rate limit restrictions on how many times we
can download an image. Let's just download (and cache) the image once, and
change most of our base images to this oci image, instead of downloading it
from the internet.

Note that we leave one docker import in there, because we do actually need
to test the code that does the docker:// import. However, the rate limit
seems to be 100 pulls/6 hours, which means you'd have to do 50 test suite
runs in six hours to hit it.

This also has an advantage of making the test suite *much* faster. We could
probably speed things up even more by fixing up places with multiple copies
around, but for now I just did a straight web -> oci transition.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>