overlay: move to mounting rootfs using LXC
The problem here is that when running unprivileged, we need to be root in
a user namespace in order to mount overlayfs, which stacker itself will not
be. However, LXC is in this position and has overlay support, so we can use
it to mount the rootfs instead.

The prep work happened in previous patches, so this is mostly about ripping
our mount detection code out (and slightly re-working the way tar
extraction works).

As a bonus, this lets us get rid of a lot of code.

However, we can no longer implement the flag --leave-unladen with this
approach, since the rootfs exists entirely in a user namespace and can't be
injected into the host's namespace.

I don't think this is a problem, since this flag is mostly unused by users.
However, it is used throughout the test suite, so that needed a bit of
refactoring.

Finally, this also breaks a bit of behavior we were using in a few places
where we were doing from type: tar, and importing it from a previous
rootfs. However, (and this already worked), we can actually use stacker://
to source these things, so people can migrate to that.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>