base layers: drop support for "type: scratch"
This never really worked that well and wasn't useful; the documentation for
this describes how obtuse it really was. Let's just drop it.

Fixes: #26

In deciding to put on this, we decided that we want to require stacker to
require a reasonable rootfs (in part due to complications with static
linking a golang binary against liblxc, causing hangs in setgroups()), so
let's add a note about the fact that we're expecting a reasonable rootfs.

We could fix this in the future by using something like the go embedding
support in 1.16 to embed a file that we can then emit and bind mount into
the container to do things like CP if we want. It would need to be a
separate statically linked binary, but that wouldn't bee too hard.

In any case, for now we punt and add these notices.

Closes: #113

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>