teach btrfs about -squashfs suffix in base layers
83e2bc23742b ("build: allow for multiple outputs") changed squashfs layers
to be output with a -squashfs suffix. If we later have a type: built layer
that wants to use this output as its base, it should be sure to append
-squashfs to get the right base layers.

This could also have resulted in mixed tar/squashfs images when using the
btrfs backend, which we have heard of reports of.

Fixes #115

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>