ensure build only base type tars work
A build only base type tar means that there will be nothing in the output,
so there's nothing to copy over, and we'll have to generate everything from
scratch ourselves.

The way this happens is slightly different between btrfs backends (it
generates one big blob) vs. overlay (one blob for the base layer and one
blob for the delta), so we need two different tests to check the results.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>