recursive-build, publish: narrow regex for stacker files
This regex will match anything containing stacker.yaml, including vim swap
files or things like test_stacker.yaml, which wouldn't be built by the
default "build" command.

Let's change this to match paths that end in /stacker.yaml, which is
probably as close semantically as we can get to the default build command.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>