overlay: run layer conversion in a userns too
When using the overlay backend, individual layers are converted from
tar->squashfs when needed (in contrast with btrfs which converts the
underlying image wholesale, preventing sharing of bits). However, this
conversion happened in the host namespace, resulting in invalid images when
used in unprivileged mode.

Let's run this conversion inside a user namespace so we generate valid
images when converting them in this way.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>