overlay: implement Delete() in a userns
Delete() actually has to walk the overlay dir and unlink() everything,
meaning that if some file/dir was not writable/owned by the user in the
host ns (e.g. a 400 file, in a 10000 uid owned dir), the delete would fail.

Let's do the delete in a userns, so it succeeds.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>