base: check the return value of storage.Delete()
When (re-)creating storage for a particular target, we previously ignored
the return value of Delete() because "hey, ENOENT is ok". But there can be
more errors than ENOENT, for example EACCES in the case of the previous
patch. Let's check the return value, and only ignore the errors that are
safe to ignore.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>