exe: run all commands in a userns
Various operations stacker can do (e.g. clean) rely on modifying/deleting
files which are potentially owned by users in the current user's subuid
allocation, which cannot be modified outside of a user namespace.

To fix this, let's run all operations inside a user namespace, so that root
in that userns can do whatever modifications it needs to do.

Since we're now running chroot in a userns, we need to propagate stdin to
the inner execution, hence that chunk of the diff.

We can also drop the wrapper testsuite-check-overlay/check-overlay wrapper,
since things are automatically run in a userns for us now.

Closes: #116
Closes: #111

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>