cache: don't fail on cache json incompatibilities
If we try to parse both the cache and the version all at once, we may fail
if some of the json representation of the cache fails.

To fix this, let's parse *just* the version number, and then if those match
continue on to parsing the rest of the cache.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>