tests: add an unprivileged axis to the test matrix
Rather than have some tests that we run unprivileged, and some that we run
privileged, let's run everything in both configurations so we can increase
coverage.

There are a bunch of changes to the testsuite needed; this is because the
testsuite runs as root, but we want to run stacker as non-root. So e.g.
when we do mkdirs or whatever for things stacker needs to be able to
read/write, we need to chown those in non-root mode.

This commit introduces the run_as helper, which runs things as the right
user in the test suite. Probably some more of the test suite modifications
could be ported to use this helper.

This commit also changes the way Clean() works slightly: the common clean
code no longer deletes (and thus no longer fails on deleting) the roots
directory. This is because in unpriv+btrfs mode, one cannot necessarily
unmount the roots dir if someone has done unpriv-setup. So we just exit(1)
with a warning in this case. The overlay code already deleted this dir, so
the extra delete code was redundant.

Finally, this commit also adds a few missing errors.Wrapf()s to errors
coming from the filesystem.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>